+++ HymsonMES/src/Libraries/Hymson.MES.Services/Dtos/Integrated/ImportBasicDataDto.cs
@@ -298,7 +298,7 @@ namespace Hymson.MES.Services.Dtos.Integrated
         /// 批次大小(去工作重心表中匹配)
         /// </summary>
         [EpplusTableColumn(Header = "批次大小(必填)", Order = 4)]
-        public int Batch { get; set; }
+        public decimal Batch { get; set; }
 
         /// <summary>
         /// 采购类型
